requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pytesseract
Pillow
requests
groq
pdf2image
